Fix wrong placed waiting room list popover
If the waiting room list is in an open state but hidden and a new participant enters the waiting room,
the waiting room list popover was opened on the top left corner.
This MR fixes this by clearing the anchor element when no more participants are in the waiting room.

Closes #965