Emoji picker upgrade
This commit contains major version upgrade of emoji-picker-react
library. (3.6.2 -> 4.4.7)

With this version, we gained the ability to lazy load emojis and more
component customizations via CSS variables. The new version also allows us
to cap the emoji version, which we did to 11 as it is the most widely supported.

With the limit of the emoji version, we fixed issue #840.

Since we cannot access the underlying input field of the emoji picker
we had to attach a focus and blur event listeners to the wrapper component
and control shortcut triggers. Shortcuts are only disallowed when
the search input field is in the focus.

This enhancement fixes issue #973.

With this commit, we also translated emoji categories to German.

Closes #840 #973