feat: show room blocker for a full room
If participant is trying to join a full room, the WebSocket opens and
controller sends a `JoinBlocked` message. This set the client into
`ConnectionState.Blocked` state. In this state the UI shows a
`RoomLoadingView` with respective text, Additionally client tries
to rejoin. Every 10 seconds it sends a `Join` signal to the
controller. If someone left the room and it's not full again,
controller will let the aforementioned participant in.

Closes #1074