Enhanced experience on poll temporary saving.
Prior to this commit, we allowed a user to be
on the "Create Poll" view after hitting the
"Save" button to temporarily save the poll.
This led to the experience where users could
hit the save button with exactly the same
details and have multiple polls stored.

This is fine, and we allow users to create
polls with the same topic and answers but we
don't really want to allow them to accidentally
create multiple polls.

Therefore, this commit brings a patch that closes
"Create" form and forces the user to explicitly
create another poll with exact details if that
is something they would want (maybe bringing
copy/clone feature there).

Closes #1109