Changed legal vote popover for users.
From this commit on, unselected participants no
longer see "Save" button as well as the token field
at the end of voting. Also, message under the key
'legal-vote-not-selected' has been moved under the
options to match successfully voted message position.

For the selected participants, "Save" button is
moved to the right side.

Closes #1071