Changed minimum duration on Poll timer to 1 minute
We are changing minimum poll duration to 1 in
order to skip formik check where we prevent duration
that is less than one minute.

We are changing timer minimum value to 1 to also
avoid formik validation.

We are changing coffee break minimum duration to 1
to prevent creation of stopwatch when the value in
the custom input is 0.

Closes #1107