Fix unresponsive hang up button
When multiple participants are leaving the meeting
simultaneously, some users experienced unresponsive
hang up button. This was caused by frequent component
re-rendering on participant leave event causing
a number of participants to change.

This commit fixes this issue by making sure that
the button will not re-render in such cases.

Closes #1145