Talking stick feature
- Added a new file called EnterpriseProvider that uses React context
to manage state and follows the Factory design pattern. This provider
renders either AutomodProvider or LegalVoteProvider based on the
specific needs of the application.
- Added a new utility function that utilizes lodash.shuffle, which
returns a new array with the same values as the original, but in a
random order. The function further ensures that the returned values are
different than the original.
- Implemented initial functionality to disable tabs within the
`MeetingSidebar` when a particular feature is active.
- Included an additional selector to associate participants with IDs
included in the talking stick feature.
- Added translations for strings related to the talking stick feature.
- New functionality was added to the project that automatically mutes
all participants except for the one who initiated the feature while it
is running.
- The import statements in the project were updated to point to the
`common` package instead. Additionally, deprecated interfaces were
removed from the monorepo.
- The `SortPopover` component is now obsolete and has been replaced by
a newer `SortPopoverMenu` component.
- A new redux selector was created to retrieve a list of all
participants, with the current user being placed at the top of the list.
- Separated participant sorting and filtering methods.

Closes #1117