Fix both coffee break and timer tabs being disabled
We should only disable the non-active tab when the other one starts -
if timer is started we disable "coffee break" tab or vice versa.

This change makes it depend on the tab key instead of the module key,
since both timer and coffee break are linked to the same module.