Fix: Poll/Voting User not able to submit their poll
After taking a legal vote and creating a new poll vote, the user who once voted on LegalVote was unable to vote
on poll voting, because in the VoteResult component we were only chacking if the user voted for LegalVote.

This MR fixes that by adding voteType to the VoteResult component and based on that chaking if the user has
voted based on the current type.

Also fixes the pop over display when starting a Poll but LegalVote result table is still in the view.

Close #1216