extend hot reload functionality for packages
Hot reload functionality is extended for `@opentalk/common`,
`@opentalk/rest-api-rtk-query` and `@opentalk/react-redux-appauth` packages.

To test it, create and modify `packages.tsconfig.json` as described in `Instructions.md`.
Start application using `yarn start:hot` and make changes in the source code of
one of the aforementioned packages.

Observe instant reload of the application with applied changes.

Closes #1294