fix audio subscription and closing screen share
The participant media state was not updated as we used the wrong
'entityAdapter' reducers in 'mediaSubscriberSlice', e.g.
'addOne()' will skip it the entity exists.
So we now just have an 'updated' case which handles all updates.
(see issue #1315)

When closing a subscriber connection we need to distinguish between
not needing the media steam and removing entirely because the publisher
terminated its connection.
(see issue #1320)