Safari: after switching audioDevice audio is not working anymore
When switching audio device while audio is running the conference would freeze or the audio will get turned off and user was not able
to turn it on. This was happening only on Safari, because safari doesn't handle automaticly all the steps for the proper device change
(eg. audio layer in Chrome dealts with devices being removed/added during ongoing audio session).

This MR adds explicit audio halt when we are calling reconfigure action for audio device change while ongoing audio session.

Closes #1201