Tweak bandwidth control for lossy networks
Originally we interpreted packet loss a symptom of a bandwidth shortage.
Recently we saw many poor WiFi networks with packet loss up to 20%.

For this case we tweak bandwidth control by adding a parameter for
the minimal video quality that is not degraded further.
This way screen share will to be always available.

Closes issue #1319