display and pin remote sharedscreen automatically
automatically switches the UI to the speaker view, if a participant starts
the screensharing. The tile of this participant is being pinned as well.
If another participant starts screen sharing meanwhile, her tile will be
pinned for others instead of the first one.
If the participant, who started screensharing as the last one, stopped
screensharing - unpin her tile for another participants.
That is, we care only about the participant, who started screensharing
as the last one.

Closes #1345