Remove login popup
To prevent the user from clicking the login button, he will be
directly navigated to the login. Weve added redux persist, to persist
the auth state will all the token information.

In addition we added an option for signing out user programaticly
from middlware (not only from the auth package hook as it was earlier).