fix functionality to set avatar default image
Avatar URL of the participant is now modified directly in the signalling API
on receive. Default image option, set by user via the environmental variable
`LIBRAVATAR_DEFAULT_IMAGE`, is appended.
After modification it is stored in the Redux store.
Default image option must not be passed to the @components anymore, therefore
it is removed from the `EnterpriseProvider`.

Closes #1401