fix guest link generation in the dashboard
A new flow for the guest link fetching in the dashboard has been implemented:
first we fetch all available invite links and check, if there is at least one
permanent guest link available. If no and current user is the creator of the
meeting, we send a request to create a permanent guest link, which is being
displayed in the meeting settings.

We don't generate a permanent link, using copy guest link option from the
meeting popover. If for some reason no permanent link is available, an error
notification will be shown to the user, asking to go to the meetings settings and
copy guest link from there.

Closes #781 #1432