improve date time picker behaviour
New reusable `DateTimePicker` is now used both in dashboard and meeting room.
It's build upon a `DateTimePicker` MUI component with two different views:
mobile and desktop.

`LocalizationProvider` is moved to the `DateTimePickerProvider` as it's
being used only for the `DateTimePicker`. Moreover it's needed there to
customize labels of the picker action buttons.

Closes #1046