Implementing ping pong control protocol with the server to check the websocket connection
This MR adds singaling connection check by sending ping echo messages on every couple of seconds.
When echo message is send we expext the same massage to be returned if singaling is working correctly.
If two echo messages are sent and no response from the signaling socket is received, the error will be thrown,
and invoked reconnect dialog.

This will prevent webrtc to continue working when signaling is down.

Clinet side will be informed pretty fast if something happened with the signaling (eg.no internet connection).

Client side will responde more agile and handle/inform the clinet and webrtc for the connection issues and act accordingly.

Closes #1554