Show Meeting Title in lobby & waiting room
This MR adds new endpoint for retriving the meeting title in rtk query, and
displays that title into lobby and waiting room.

For invited users it creates a conditional headers for this endpoint
to override the main prepare headers by adding the invite code as an
authentication.

Removes the typescript definition for RoomId, because of the conflicts with
the same deifinition inside common package, and for now just re-export it from
common package instead of creating his own definition.

Closes #1398