Fix Password notifications don't disappear
This MR fixes the problem with password notification not closing after multiple
wrong password tries by checking if the same notification exist before displaying it.
This will prevent multiple notification keys that were causing the notification to not close
after having multiple of them displayed.
Not we guranty that only one will be renreded, and on success the same one will be closed.