integrate quick guide in the app
Implements quick start guide button in the lobby and in the room. On click a popper is being shown with embedded quick guide in the SVG form. The size of the SVG is only roughly scaled to fit the most screen size. Proper responsivness will be implemented in future if desired.

A11y has been considered as well.
Escape key shall always close the popper, even if it has lost the focus.
Space key has been disabled in the room because of the push to talk feature.

The help section of dashboard has been restructured, to integrate quick  guide subsection there. Here the quick guide shall be opened in a new browser tab.

There are two versions of quick start quide: de and en. Proper language version of the quick guide shall be picked up. For this purpose useQuickStartUrl custom hook has been implemented.

Closes #1525