implement confidence threshold for virtual background
Implement confidence threshold and alpha blending of the edge area
between min and max) to smooth the sharp corners a little bit.
In such a way we have a better control of the edge area around the person mask.
Objects around the person can be hidden with slightly better precision.

Min and max coeffs has been determined during QA session

Closes #1595