Fix problem with the wrongly sorted users by first hand raised
Problem occured due to the fact that we were setting
hand raised timestamp before sending the signaling
message to the backend, this caused slight disbalance
and gap where we would see us in the different place
in the list than the other participants due to the differences
in preset and received timestamp.

Backend introduced new control message for hand_raised
which allows us to know when backend acknowledged
that we raised hand and we update that information locally
on the pong and not the ping as we used to do
making hand raise feel slight slower but much more
stable as we will no longer be out of the sync with the
backend and we will have the same information
accross the users and therefore the proper and mirrored
participant list.