fix: add aria-label on notification close button
fix: Remove naively added SnackbarProps from the notifications

As a part of https://git.opentalk.dev/opentalk/frontend/web/web-app/-/merge_requests/1217
we naively added `SnackbarProps` with the intention of overriding notistack predefined role.

This turned out to be an imposible task, at least in the actual version
of the notistack, meaning that `SnackbarProps` will change root of the
notification, that is actually not holding the role leaving us with nested
alerts and double pronunciation of the notification content with the
screenreader.

With this commit, we are going back to the implementation without any
`SnackbarProps` props being defined, leaving us with only one alert.

chore: move consent recording notification to notistack variation

Due to the fact that we cannot simply override default notification
role and aria attributes, we have to apply technique of custom
notification content by defining a new variation.

In this case, we create a variation called "consent" that can be called
through notifications.consent and is using ConsentNotification
component under the hood.

Downside of the new consent notification is that the icon is not
matching the default notistack warning icon, but this will be
investigated and resolved in the upcoming fixes.

fix: make breakout room alertdialog

Breakout room notification has been built on top of now
deprecated API by passing custom content through options.
This still works but just as a note, we need to migrate it to the
apporach from recording consent notification.

This commit makes breakout room notification `alertdialog`
along with the aria-describedby pointing to the title of the
notification.

In this commit, we introduced two new
notification variations that covers talking stick
unmuted and muted states.

chore: add more menu unit tests