fix: wobbling meeting timer
Issue was caused by text-align: center CSS property which tried
to maintain center position with different width characters
causing whole timer to shake.

With this commit, we switched to text-align: left and adjusted
container size to save some space.

This commit also introduces QA helper so they don't
wait for an hour to confirm that longer times don't
wobble around. All they have to do is to add devMode
in the local storage and to set window.DEBUG = true
in the console.